metal adapt

+++ cocos/core/pipeline/render-pipeline.ts
@@ -910,7 +910,7 @@ export abstract class RenderPipeline {
         this._quadVB.update(verts);
 
         // create index buffer
-        const ibStride = Uint8Array.BYTES_PER_ELEMENT;
+        const ibStride = Uint16Array.BYTES_PER_ELEMENT;
         const ibSize = ibStride * 6;
 
         this._quadIB = this._device.createBuffer({
@@ -924,7 +924,7 @@ export abstract class RenderPipeline {
             return false;
         }
 
-        const indices = new Uint8Array(6);
+        const indices = new Uint16Array(6);
         indices[0] = 0; indices[1] = 1; indices[2] = 2;
         indices[3] = 1; indices[4] = 3; indices[5] = 2;
