+++ cocos/core/splash-screen.ts
@@ -347,7 +347,7 @@ export class SplashScreen {
         this.textVB.update(verts);
 
         // create index buffer
-        const ibStride = Uint8Array.BYTES_PER_ELEMENT;
+        const ibStride = Uint16Array.BYTES_PER_ELEMENT;
         const ibSize = ibStride * 6;
 
         this.textIB = this.device.createBuffer({
@@ -357,7 +357,7 @@ export class SplashScreen {
             stride: ibStride,
         });
 
-        const indices = new Uint8Array(6);
+        const indices = new Uint16Array(6);
         indices[0] = 0; indices[1] = 1; indices[2] = 2;
         indices[3] = 1; indices[4] = 3; indices[5] = 2;
         this.textIB.update(indices);
@@ -430,7 +430,7 @@ export class SplashScreen {
         this.vertexBuffers.update(verts);
 
         // create index buffer
-        const ibStride = Uint8Array.BYTES_PER_ELEMENT;
+        const ibStride = Uint16Array.BYTES_PER_ELEMENT;
         const ibSize = ibStride * 6;
 
         this.indicesBuffers = device.createBuffer({
@@ -440,7 +440,7 @@ export class SplashScreen {
             stride: ibStride,
         });
 
-        const indices = new Uint8Array(6);
+        const indices = new Uint16Array(6);
         indices[0] = 0; indices[1] = 1; indices[2] = 2;
         indices[3] = 1; indices[4] = 3; indices[5] = 2;
         this.indicesBuffers.update(indices);
